Add a GraphQLPlaygroundApp
