+++ .changeset/smooth-toys-joke/changes.md
@@ -1,0 +1,1 @@
+First iteration of the GraphQLPlaygroundApp.
