+++ demo-projects/meetup/package.json
