+++ packages/app-graphql-playground/.npmignore
@@ -1,0 +1,2 @@
+**/*.md
+**/*.test.js
