+++ packages/app-graphql-playground/CHANGELOG.md
@@ -1,0 +1,1 @@
+# @keystone-alpha/app-graphql-playground
