+++ packages/app-graphql-playground/lib/devQuery.js
