+++ packages/app-graphql-playground/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "@keystone-alpha/app-graphql-playground",
+  "description": "KeystoneJS GraphQL Playground App.",
+  "version": "0.0.0",
+  "author": "The KeystoneJS Development Team",
+  "license": "MIT",
+  "engines": {
+    "node": ">=8.4.0"
+  },
+  "dependencies": {
+    "@apollographql/graphql-playground-html": "^1.6.24",
+    "@apollographql/graphql-playground-react": "^1.7.31",
+    "@keystone-alpha/logger": "^2.0.1",
+    "@keystone-alpha/session": "^2.0.1",
+    "@keystone-alpha/utils": "^3.0.2",
+    "body-parser": "^1.18.2",
+    "chalk": "^2.4.2",
+    "falsey": "^1.0.0",
+    "graphql-tag": "^2.10.1",
+    "object-hash": "^1.3.1",
+    "terminal-link": "^1.3.0"
+  },
+  "peerDependencies": {
+    "express": "^4.17.1"
+  }
+}
