+++ packages/app-graphql/README.md
@@ -36,10 +36,9 @@ module.exports = {
 
 ## Config
 
-| Option         | Type     | Default                               | Description                                                                                          |
-| -------------- | -------- | ------------------------------------- | ---------------------------------------------------------------------------------------------------- |
-| `cors`         | `Object` | `{ origin: true, credentials: true }` | Options passed directly to the express [cors](https://github.com/expressjs/cors) middleware package. |
-| `apiPath`      | `String` | `/admin/api`                          | Change the API path                                                                                  |
-| `graphiqlPath` | `String` | `/admin/graphiql`                     | Change the Apollo GraphQL playground path                                                            |
-| `schemaName`   | `String` | `admin`                               | Change the graphQL schema name (not recommended)                                                     |
-| `apollo`       | `Object` | `{}`                                  | Options passed directly to Apollo Server                                                             |
+| Option         | Type     | Default           | Description                                      |
+| -------------- | -------- | ----------------- | ------------------------------------------------ |
+| `apiPath`      | `String` | `/admin/api`      | Change the API path                              |
+| `graphiqlPath` | `String` | `/admin/graphiql` | Change the Apollo GraphQL playground path        |
+| `schemaName`   | `String` | `admin`           | Change the graphQL schema name (not recommended) |
+| `apollo`       | `Object` | `{}`              | Options passed directly to Apollo Server         |
