+++ packages/app-graphql/package.json
@@ -8,25 +8,18 @@
     "node": ">=8.4.0"
   },
   "dependencies": {
-    "@apollographql/graphql-playground-html": "^1.6.24",
-    "@apollographql/graphql-playground-react": "^1.7.31",
     "@keystone-alpha/logger": "^2.0.1",
     "@keystone-alpha/session": "^2.0.1",
+    "@keystone-alpha/app-graphql-playground": "^0.0.0",
     "@keystone-alpha/utils": "^3.0.2",
     "apollo-errors": "^1.9.0",
     "apollo-server-express": "^2.9.1",
-    "body-parser": "^1.18.2",
-    "chalk": "^2.4.2",
     "cuid": "^2.1.6",
     "ensure-error": "^1.0.0",
-    "falsey": "^1.0.0",
     "graphql": "^14.4.2",
-    "graphql-tag": "^2.10.1",
     "nanoassert": "^2.0.0",
-    "object-hash": "^1.3.1",
     "serialize-error": "^3.0.0",
-    "stack-utils": "^1.0.2",
-    "terminal-link": "^1.3.0"
+    "stack-utils": "^1.0.2"
   },
   "peerDependencies": {
     "express": "^4.17.1"
