+++ packages/session/package.json
